Generalize TORCH_CUDA_ARCH_LIST warning suppression for NNCF extensions
